# api.yaml
openapi: 3.0.3
info:
  title: Placement Portal Application (PPA V5) API
  version: 1.0.0
  description: |
    RESTful API for a campus placement management system.
    Supports three roles: Admin, Company, and Student.
servers:
  - url: http://localhost:5001
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentProfile:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        full_name:
          type: string
        department:
          type: string
        cgpa:
          type: number
        graduation_year:
          type: integer
        resume_path:
          type: string
        phone:
          type: string
        bio:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        is_blacklisted:
          type: boolean
        skills:
          type: array
          items:
            type: string
...
    PlacementDrive:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        company_name:
          type: string
        drive_name:
          type: string
        job_title:
          type: string
        job_description:
          type: string
        eligibility_branch:
          type: string
        min_cgpa:
          type: number
        eligible_year:
          type: integer
        application_deadline:
          type: string
          format: date-time
        location:
          type: string
        salary:
          type: string
        job_type:
          type: string
          enum: [Full-time, Internship, Contract]
        status:
          type: string
          enum: [pending, approved, rejected, closed]
        created_at:
          type: string
          format: date-time
        application_count:
          type: integer
    Application:
      type: object
      properties:
        id:
          type: integer
        student_id:
          type: integer
        drive_id:
          type: integer
        student_name:
          type: string
        student_email:
          type: string
        student_cgpa:
          type: number
        student_department:
          type: string
        drive_name:
          type: string
        company_name:
          type: string
        job_title:
          type: string
        application_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected, withdrawn]
        interview_type:
          type: string
        remarks:
          type: string
        cover_letter:
          type: string
    Interview:
      type: object
      properties:
        id:
          type: integer
        application_id:
          type: integer
        interview_date:
          type: string
          format: date-time
        mode:
          type: string
          enum: [Online, Offline]
        venue:
          type: string
        result:
          type: string
          enum: [pending, passed, failed]
...
    Message:
      type: object
      properties:
        message:
          type: string
paths:
...
  /api/students/register:
    post:
      tags: [Student]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, email, password]
              properties:
                full_name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                department:
                  type: string
                cgpa:
                  type: number
                graduation_year:
                  type: integer
                phone:
                  type: string
      responses:
        "201":
          description: Student registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/students/profile:
    get:
      tags: [Student]
      summary: Get student profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Student profile data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
    put:
      tags: [Student]
      summary: Update student profile
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                department:
                  type: string
                cgpa:
                  type: number
                graduation_year:
                  type: integer
                phone:
                  type: string
                bio:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  student:
                    $ref: "#/components/schemas/StudentProfile"
  /api/students/upload-resume:
    post:
      tags: [Student]
      summary: Upload resume (PDF only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                resume:
                  type: string
                  format: binary
      responses:
        "200":
          description: Resume uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/student/drives:
    get:
      tags: [Student]
      summary: Browse approved placement drives
      security:
        - BearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: branch
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of approved drives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlacementDrive"
  /api/student/drives/{id}:
    get:
      tags: [Student]
      summary: Get drive details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Drive detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacementDrive"
  /api/student/apply/{drive_id}:
    post:
      tags: [Student]
      summary: Apply to a placement drive
      security:
        - BearerAuth: []
      parameters:
        - name: drive_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cover_letter:
                  type: string
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/student/applications:
    get:
      tags: [Student]
      summary: Get student's applications
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
  /api/student/applications/{id}/withdraw:
    put:
      tags: [Student]
      summary: Withdraw an application
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /api/student/history:
    get:
      tags: [Student]
      summary: Get placement history
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Placement history records
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /api/student/interviews:
    get:
      tags: [Student]
      summary: Get scheduled interviews
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
  /api/student/export-applications:
    post:
      tags: [Student]
      summary: Export applications as CSV (async Celery task)
      security:
        - BearerAuth: []
      responses:
        "202":
          description: Export job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  job_id:
                    type: integer
  /api/student/download-export/{job_id}:
    get:
      tags: [Student]
      summary: Download exported CSV file
      security:
        - BearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/student/ats-check:
    post:
      tags: [Student]
      summary: Run ATS resume compatibility check against a drive
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drive_id]
              properties:
                drive_id:
                  type: integer
      responses:
        "200":
          description: ATS analysis result
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: object
                    properties:
                      score:
                        type: number
                      analysis:
                        type: string